requirement.txt
absl-py==0.13.0
appdirs==1.4.4
asgiref==3.3.1
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
cycler==0.10.0
distlib==0.3.1
Django==3.1.4
filelock==3.0.12
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
flatbuffers==1.12
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.1
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.3
metrics==0.3.3
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.3.2
pathlib2==2.3.6
pathspec==0.5.5
Pillow==8.3.1
pip-install-privates==0.6.0
pip-install-test==0.5
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.2.0
PyMySQL==0.10.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2020.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.2.0
typing-extensions==3.7.4.3
urllib3==1.26.6
virtualenv==20.2.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
